src/requirements.txt
aioredis==1.1.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
boto3==1.7.58
botocore==1.10.58
certifi==2018.4.16
cffi==1.11.5
channels==2.1.2
channels-redis==2.2.1
chardet==3.0.4
click==6.7
constantly==15.1.0
croniter==0.3.25
cryptography==2.3
daphne==2.2.1
Django==2.0.4
django-countries==5.3.1
django-rq==1.2.0
django-timezone-field==3.0
docutils==0.14
ecdsa==0.13
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jmespath==0.9.3
msgpack==0.5.6
natsort==5.2.0
Pillow==5.1.0
psycopg2==2.7.5
pycparser==2.18
PyHamcrest==1.9.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
raven==6.9.0
redis==2.10.6
requests==2.19.1
rq==0.12.0
rq-scheduler==0.8.3
s3transfer==0.1.13
six==1.11.0
sshpubkeys==3.1.0
Twisted==18.7.0
txaio==2.10.0
urllib3==1.23
uWSGI==2.0.17.1
zope.interface==4.5.0
